P4-3091 renaming package 'commands' to 'cli' for a clearer term of what it is meant for.
